Chamados - BOWSE e ABRIR CHAMADO
Related Work Items: #37, #38

+++ Sindemed/Models/Repositories/AssociadoViewModel.cs
@@ -147,7 +147,7 @@ namespace Sindemed.Models.Repositories
         public string email3 { get; set; }
 
         [DisplayName("Usuário")]
-        public Nullable<decimal> usuarioId { get; set; }
+        public Nullable<int> usuarioId { get; set; }
 
         [DisplayName("Observação")]
         public string observacao { get; set; }
